Allow settings paramter to take preset value
Currently, we can provide either a PowerShell data file or a
hashtable to the settings parameter. But we would like users to
take advantage of the builtin setting files {PSGallery,DSC,etc}
by providing only their names as values. This provides a easier
way to use these builtin setting files.